persephone: make clevis jwe get copied to store
i hit a weird issue where old generations were referencing this through a path to the flake in the store, but i guess that path got garbage-collected. this prevented installing the boot loader until those old generations were deleted. it also happened after i accidentally filled up my disk: not sure if that was related or just caused me to GC.

i don't know if this will fix that problem or not. worth a try.