music-assistant: make airplay actually work
package cliraop and build it through nix. this is not strictly necessary for it to work: the previous patchelf nonsense was working, but it was needed to be able to debug it, and i prefer it to using a bundled binary.

opening the udp ports on the other hand is what is necessary