fix deadlock with spiffe-creds and vault login
the spiffe-creds service needs to live long enough that it can survive
the life of any individual Exec* step in a service that uses it.
otherwise it will exit, and then need to be restarted and try to login
to vault again. this seems to cause a deadlock with systemd, as the
login to vault requires getting SPIFFE creds, which then needs to talk
to systemd to determine which service is asking for them.

the solution is to let the service live longer, and update the vault
token in the background.